Day 6: Add part #2
The use of an array worked until a certain scale. After that, the previous implementation were hit memory issues. To fix it, the simple solution was to think of all the fish as a pool rather than individual fish, and count the number of fish for each timer, reducing the inner loop of just 9 iteration.